+++ src/hardware/mame/sn76496.cpp
@@ -279,7 +279,7 @@ void sn76496_base_device::device_start()
 	{
 		// limit volume to avoid clipping
 		if (out > MAX_OUTPUT / 4) m_vol_table[i] = MAX_OUTPUT / 4;
-		else m_vol_table[i] = out;
+		else m_vol_table[i] = (int32_t)out;
 
 		out /= 1.258925412; /* = 10 ^ (2/20) = 2dB */
 	}
